+++ demos/client-example-server/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @prosopo/client-example-server
 
+## 2.7.21
+### Patch Changes
+
+- 1f3a02f: Release 3.1.8
+- Updated dependencies [1f3a02f]
+  - @prosopo/server@2.9.17
+
 ## 2.7.20
 ### Patch Changes