+++ dev/lint/CHANGELOG.md
@@ -1,5 +1,10 @@
 # @prosopo/lint
 
+## 2.7.10
+### Patch Changes
+
+- 1f3a02f: Release 3.1.8
+
 ## 2.7.9
 ### Patch Changes